fix(runtime): nv_box_int keeps all 64 bits
It cast the incoming i64 to int32 before boxing, so a literal above 2^31 was already
wrong by the time it reached the object.

+++ src/backend/runtime/nir_bridge.c
@@ -51,7 +51,7 @@ static int obj_is_truthy(NvObject* obj) {
 
 NvObject* nv_box_int(int64_t v) {
     Value out = {NULL};
-    create_int(&out, (int32_t)v);
+    create_int(&out, v);
     return out.obj;
 }
